COFFEE-019-fixed test, more translations-[GA, FA]
